v7.2.0: rename to 'AI Sync for Jekyll & GitHub Pages' + WP.org compliance
Renames to comply with WP.org trademark guidelines (display name uses 'for X' pattern to signal non-affiliation with Jekyll/GitHub projects). Slug, text-domain, and stored data unchanged — in-place upgrades work without configuration loss.

Plugin Check: 0 errors / 0 warnings (canonical wp.org check via plugin-check 2.0.0).
Activation smoke test: passes on clean WP 7.0 with WP_DEBUG=true.
phpcs WordPress security + PHPCompatibilityWP 7.4-: 0 violations.

- Plugin Name: 'Ray Bogman AI Sync for Jekyll & GitHub Pages' (+ non-affiliation line in Description)
- Updated all in-UI headers, About tab (with trademark disclaimer), FAQ blurb, footer
- readme: title, short description (<=150 chars), Description, Tags cleaned of bare brand terms, Tested up to: 7.0
- esc_attr() on inline style at class-admin.php:623
- (int) cast on strlen() output at class-articles-table.php:181
- set_time_limit(300) hardened: disable_functions check + scoped phpcs annotation
- Version bumped: 7.1.0 -> 7.2.0 (MINOR per SemVer; rename + compliance, no breaking changes)

+++ includes/class-articles-table.php
@@ -178,7 +178,7 @@ class WPJS_Articles_Table extends WP_List_Table {
 				</div>
 			</div>
 			<p class="description" style="margin:4px 0 0;">
-				<span id="wpjs-desc-count-<?php echo (int) $pid; ?>"><?php echo strlen( $desc ); ?></span>/160 characters
+				<span id="wpjs-desc-count-<?php echo (int) $pid; ?>"><?php echo (int) strlen( $desc ); ?></span>/160 characters
 			</p>
 
 			<?php if ( $images ) : ?>
